+++ controllers/middleware/initial.js
@@ -15,7 +15,6 @@ const generic = async (req, res, next) => {
 
     // template variables
     res.locals._view = {
-        _status: { isPanel: req.url.includes('/home') },
         flash: { error: req.flash('error'), info: req.flash('info'), pass: req.flash('pass') },
         title: res.locals._site.title,
         user: req.session.user,