+++ controllers/routers/account.js
@@ -17,7 +17,7 @@ const { userModel } = require('../../models/');
 exports.account.signup = {
     get: (req, res) => {
         if (req.isAuthenticated()) return res.redirect('/home');
-        return res.render('./root/account/signup');
+        return res.render('./__root__/account/signup');
     },
     post: [_M_.passwordValidation, async (req, res) => {
         const newUser = await userModel.register(new userModel(req.body), req.body.password.new);
@@ -32,7 +32,7 @@ exports.account.signup = {
 exports.account.signin = {
     get: (req, res) => {
         if (req.isAuthenticated()) return res.redirect('/home');
-        return res.render('./root/account/signin');
+        return res.render('./__root__/account/signin');
     },
     post: (req, res, next) => {
         if (req.isAuthenticated()) return res.redirect('/home');
@@ -41,7 +41,7 @@ exports.account.signin = {
             if (!authUser) return next(new ClientError(20002));
             return req.logIn(authUser, err => {
                 if (err) return next(err); // todo: throw an error
-                authUser.UpdateSignInLog();
+                authUser.updateLastTimeLog('signIn');
                 req.session.cookie.expires = req.body.isPersisted ? new Date(Date.now() + 14 * 24 * 3600000) : false;
                 req.session.user = { _id: authUser._id, picture: authUser.picture, nickname: authUser.nickname };
                 req.flash('info', `Welcome back ${authUser.nickname}`);