+++ controllers/routers/admin.js
@@ -14,11 +14,11 @@ const { configModel, mediaModel, userModel } = require('../../models/');
 //  CONTROLLERS
 // ==============================
 home.main = {
-    get: (req, res) => res.render('./root/'),
+    get: (req, res) => res.render('./__root__/'),
 };
 
 home.setting = {
-    get: (req, res) => res.render('./root/setting'),
+    get: (req, res) => res.render('./__root__/setting'),
     patch: async (req, res) => {
         await configModel.updateSettings(req.body.configs);
         return res.redirect('back');
@@ -28,7 +28,7 @@ home.setting = {
 home.profile = {
     get: (req, res) => {
         res.locals._view.user = req.user._doc;
-        return res.render('./root/account/profile');
+        return res.render('./__root__/account/profile/profile');
     },
     patch: async (req, res) => {
         const raw = { info: req.body.profile.info, nickname: req.body.profile.nickname };
@@ -42,21 +42,21 @@ home.profile = {
 home.profile_editor = {
     get: (req, res) => {
         res.locals._view.user = req.user._doc;
-        return res.render('./root/account/profile_editor');
+        return res.render('./__root__/account/profile/editor');
     },
 };
 
 home.security = {
-    get: (req, res) => res.render('./root/account/security'),
+    get: (req, res) => res.render('./__root__/account/security'),
     patch: [_M_.passwordValidation, async (req, res) => {
         await req.user.changePassword(req.body.password.old, req.body.password.new);
         req.flash('info', 'Password have been successfully changed.');
-        return res.redirect('back');
+        return res.redirect('/home/profile');
     }],
 };
 
 home.upload = {   // todo: to be integrated in profile and media manager
-    get: (req, res) => res.render('./root/upload'),
+    get: (req, res) => res.render('./__root__/upload'),
     post: [_M_.hireBusboy({ fileSize: 25*1048576 }), async (req, res) => {
         if (req.body.busboySlip.mes.length > 0) req.body.busboySlip.mes.forEach(mes => req.flash('error', mes));
         if (req.body.busboySlip.raw.length > 0) {